+++ src/Fantomas.Core/SyntaxOak.fs
@@ -502,11 +502,16 @@ type PatParenNode(openingParen: SingleTextNode, pat: Pattern, closingParen: Sing
     member val Pattern = pat
     member val ClosingParen = closingParen
 
-type PatTupleNode(pats: Pattern list, range) =
+type PatTupleNode(items: Choice<Pattern, SingleTextNode> list, range) =
     inherit NodeBase(range)
 
-    override val Children: Node array = [| yield! (List.map Pattern.Node pats) |]
-    member val Patterns = pats
+    override val Children: Node array =
+        [| for item in items do
+               match item with
+               | Choice1Of2 p -> Pattern.Node p
+               | Choice2Of2 comma -> comma |]
+
+    member val Items = items
 
 type PatStructTupleNode(pats: Pattern list, range) =
     inherit NodeBase(range)